Activated save button fucntionality
