+++ index.html
@@ -49,7 +49,7 @@
             <dialog class="dialog" id="park"></dialog>
         
         </section>
-        <input type="submit" value="Save Itinerary">
+        <input type="submit" value="Save Itinerary" id="itinSaveButton">
     </div>
     <div>
     <section>
