+++ scripts/main.js
@@ -1,16 +1,17 @@
-import { useDb } from "./save/getDataFromDom.js";
+// import { useDb } from "./save/getDataFromDom.js";
 import {sendDropDownToDom} from "./nav/dropDown.js";
 import {SendNavToDom} from "./nav/nav.js";
 import {openNav} from "./nav/nav.js";
 import {closeNav} from "./nav/nav.js";
+import {saveButton} from "./save/writeToDb.js";
 sendDropDownToDom()
 SendNavToDom()
 openNav()
 closeNav()
-useDb();
-
-
-
+const saveButtonId = document.getElementById("itinSaveButton");
+saveButtonId.onclick = () =>{
+    saveButton();
+}
 
 const modal = document.getElementById("eateryContent")
 const button = document.getElementById("eateryButton")
