+++ scripts/save/getDataFromDom.js
@@ -3,14 +3,7 @@ import { getParksObject } from "./parksToJson.js";
 import { getAttractionObject } from "./attractionToJson.js";
 import { getEateryObject } from "./eateryToJson.js";
 
-// Itinerary object
-export const  itinObject = {
-    // TODO: hook up data in getter functions to retreive data from dom.
-    eatery:getEateryObject(),
-    attraction:getAttractionObject(),
-    parks:getParksObject(),
-    weather:getWeatherObj()
-}
+
 // Fetch call to POST itinerary to JSON server
 const sendToDataBSe = obj => {
     return fetch("http://localhost:8088/itineraries", {
@@ -23,6 +16,6 @@ const sendToDataBSe = obj => {
         .then(response => response.json())
 }
 // Fucntion to be called by when save button is clicked.
-export const useDb = () => {
+export const useDb = (itinObject) => {
     sendToDataBSe( itinObject);
 }
